requirements.txt
tensorflow<=1.15.0,>=1.14.0
tensorflow-gpu<=1.15.0,>=1.14.0
tensorflow_hub==0.6.0
bert-tensorflow==1.0.1
numpy==1.19.5
pytest==5.3.2
rusenttokenize==0.0.5
nltk
tqdm==4.60.0
hydra-core
omegaconf